README2.txt
Any coders out there that are good with iOS, newer versions that can help me identify some os coding on suspected hacked devices?!

I’ve been suspecting some foul play on my devices and think my system has been high jacked, including iPhone 12 Pro Max, Apple Watch Series 4 and iPad 9th generation. I think they got in via BLE and got ahold of the ports and from there got control of the whole system. I notice phishing pages for my apps that I’m usually logged into that will go straight to sign in page and I will hit cancel and the sign in page will drop and I will be logged in as It should be. I’ve noticed static url’s with html that has different coding kits that are not normally there. I’ve tried to view my open ports and it will not show. I’ve requested logs from the app Surge and the DNS’s that show are static and give back no response. And biggest of all besides the glitches in the UI is the analytical data. Which I know is for apple and app devs for when apps or programs crash or present issues, and should have a few different logs and codes in it, however mine on a replacement new phone that’s only a few months old should only have ten or so… My devices are filled with them and I can show whoever can help examples of everything. I’m not a professional, I do have a degree in computer science but have never messed with mobile devices until recently while trying to figure this crap out. I have in the past learned Java script and html, but not fluently enough to fix this problem. I need to get a Mac to fight back but if I could get some help and advice, I’d super appreciate it. 
Thanks for reading.